Add remaining API functionality

+++ cgi_bin/views.py
@@ -25,8 +25,12 @@ def landing(request):
 def make_payment(request):
     if request.user.is_authenticated():
         user = request.user.bankuser
+        redir = '../show/show-user'
     elif has_admin_token(request) and User.objects.filter(username=request.GET['user_name']).exists():
         user = BankUser.objects.get(user=User.objects.get(username=request.GET['user_name']))
+        redir = '../show/show-user?user_name=' + user.user.username
+    else:
+        return HttpResponseRedirect('../login')
     if request.REQUEST['other_party'] and request.REQUEST['amount']:
         dest_user = get_object_or_404(User, username=request.REQUEST['other_party'])
         amount = int(request.REQUEST['amount'])
@@ -41,24 +45,25 @@ def make_payment(request):
                     dest_user.bankuser.save()
                     t2 = Transaction(user=dest_user.bankuser, credit=amount, typeof=user, balance=dest_user.bankuser.balance)
                     t2.save()
-                    return HttpResponseRedirect('../show_user')
+                    return HttpResponseRedirect(redir)
                 else:
                     return HttpResponse("You do not have enough money to complete this transfer.")
             else:
                 return HttpResponse("The target user was not found.")
         else:
             return HttpResponse("You cannot request this kind of transfer.")
-    else:
-        return HttpResponse("You are not logged in.")
 
 def show_user(request):
-    if has_admin_token(request) and request.GET['user_name']:
-        result = BankUser.objects.get(user=User.objects.get(username=request.GET['user_name']))
-    else:
-        result = request.user.bankuser
-    transaction_history = Transaction.objects.filter(user=result).order_by('pk')
-    context = {'transaction_history': transaction_history, 'user' : result}
-    return render(request, 'cgi_bin/show-user.html', context)
+    if has_admin_token(request) or request.user.is_authenticated():
+        if has_admin_token(request) and request.GET['user_name']:
+            result = BankUser.objects.get(user=User.objects.get(username=request.GET['user_name']))
+        elif request.user.is_authenticated():
+            result = request.user.bankuser
+        transaction_history = Transaction.objects.filter(user=result).order_by('pk')
+        context = {'transaction_history': transaction_history, 'user' : result}
+        return render(request, 'cgi_bin/show-user.html', context)
+    return HttpResponseRedirect('../login')
+
 # TODO: Only allow the RDP server to connect
 def admintoken(request):
     if True: # request.meta['REMOTE_ADDR'] == '192.168.1.4':
@@ -78,7 +83,10 @@ def admintoken(request):
             return HttpResponse("Wrong!")
 
 def has_admin_token(request):
-    return AdminSession.objects.filter(access_token=request.COOKIES['access_token']).exists()
+    try:
+        return AdminSession.objects.filter(access_token=request.COOKIES['access_token']).exists()
+    except KeyError:
+        return False
 
 def find_user(request):
    if has_admin_token(request):
@@ -89,18 +97,36 @@ def find_user(request):
            return HttpResponse("User not found!")
 
 def create_user(request):
-   newuser = User(username=request.POST['name'], password=request.POST['password'])
-   newuser.save()
-   newbankuser = BankUser(user=newuser)
-   newbankuser.save()
-   return HttpResponseRedirect("cgi-bin/show/show-user?user_name=" + newuser.username)
+   if has_admin_token(request):
+       newuser = User(username=request.POST['name'], password=request.POST['password'])
+       newuser.save()
+       newbankuser = BankUser(user=newuser)
+       newbankuser.save()
+       return HttpResponseRedirect("cgi-bin/show/show-user?user_name=" + newuser.username)
 
 def delete_user(request):
-   deluser = User(username=request.POST['user_name'])
-   if deluser:
-       delbankusesr = BankUser(user=deluser)
-       delbankuser.delete()
-       deluser.delete()
-       return HttpResponse("User " + request.POST['user_name'] + " deleted!")
-   else:
-       return HttpResponse("User " + request.POST['user_name'] + " not found!")
+   if has_admin_token(request):
+       deluser = User(username=request.POST['user_name'])
+       if deluser:
+           delbankusesr = BankUser(user=deluser)
+           delbankuser.delete()
+           deluser.delete()
+           return HttpResponse("User " + request.POST['user_name'] + " deleted!")
+       else:
+           return HttpResponse("User " + request.POST['user_name'] + " not found!")
+
+def make_deposit(request):
+    if has_admin_token(request) and request.GET['user_name'] and request.GET['amount']:
+        user = get_object_or_404(User, username=request.REQUEST['other_party'])
+        if request.GET['amount'] > 0:
+            user.bankuser.balance += request.GET['amount']
+            user.bankuser.save()
+            return HttpResponseRedirect("../show-user?user_name=" + user.username)
+
+def make_withdrawal(request):
+    if has_admin_token(request) and request.GET['user_name'] and request.GET['amount']:
+        user = get_object_or_404(User, username=request.REQUEST['other_party'])
+        if request.GET['amount'] > 0 and user.bankuser.balance - request.GET['amount'] >= 0:
+            user.bankuser.balance -= request.GET['amount']
+            user.bankuser.save()
+            return HttpResponseRedirect("../show-user?user_name=" + user.username)
